Revert "Fix incorrect defaults for io queue iops/bandwidth"
This reverts commit 54509f86caa3ad198f9f5eb8aa7fef589ed9f701.

The fix steps on an issue #2817 in a weird manner -- when calculating
the maximum length of a request, the calculating loop overflows another
value:

request size -> capacity
         512 -> 0
        1024 -> 0
        2048 -> 0
        4096 -> 0
        8192 -> 0
       16384 -> 0
       32768 -> 0
       65536 -> 0
      131072 -> 0
      262144 -> 0
      524288 -> 0
     1048576 -> 0
     2097152 -> 0
     4194304 -> 0
     8388608 -> 0
    16777216 -> 0
    33554432 -> 0
    67108864 -> 0
   134217728 -> 0
   268435456 -> 0
   536870912 -> 0
  1073741824 -> 0
 -2147483648 -> 2097151999

I.e. -- at some point it checks negative request size, renders
"infinite" capacity and bails out. That's technically correct, but it's
correct accidentally. Need a controllable behavior.

+++ include/seastar/core/io_queue.hh
@@ -181,8 +181,8 @@ public:
 
     struct config {
         unsigned id;
-        unsigned long req_count_rate = std::numeric_limits<unsigned long>::max();
-        unsigned long blocks_count_rate = std::numeric_limits<unsigned long>::max();
+        unsigned long req_count_rate = std::numeric_limits<int>::max();
+        unsigned long blocks_count_rate = std::numeric_limits<int>::max();
         unsigned disk_req_write_to_read_multiplier = read_request_base_count;
         unsigned disk_blocks_write_to_read_multiplier = read_request_base_count;
         size_t disk_read_saturation_length = std::numeric_limits<size_t>::max();
